Hide Player Inputs
Updated the player one and two inputs to the get pass function from the library getpass. This hides whatever the player is typing in as their selection for rock/paper/scissors. Previously you could see what player one was inputting and player two would always be able to win as they could see the first players selection to counter.